requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-multiselectfield==0.1.13
fontawesomefree==6.6.0
idna==3.10
pillow==11.0.0
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
